Compacting Code => So Long, Stripping Data Containers...
