+++ src/main/java/org/quiltmc/loader/impl/launch/knot/KnotClassDelegate.java
@@ -18,8 +18,7 @@ package org.quiltmc.loader.impl.launch.knot;
 
 import net.fabricmc.api.EnvType;
 
-import org.quiltmc.loader.api.plugin.solver.ModLoadOption;
-import org.quiltmc.loader.impl.transformer.PackageRequiresStrippingVisitor;
+import org.quiltmc.loader.impl.transformer.PackageStrippingData;
 import org.quiltmc.loader.impl.util.LoaderUtil;
 import org.objectweb.asm.ClassReader;
 import org.quiltmc.loader.api.ModContainer;
@@ -29,11 +28,9 @@ import org.quiltmc.loader.impl.game.GameProvider;
 import org.quiltmc.loader.impl.launch.common.QuiltCodeSource;
 import org.quiltmc.loader.impl.launch.common.QuiltLauncherBase;
 import org.quiltmc.loader.impl.patch.PatchLoader;
-import org.quiltmc.loader.impl.transformer.PackageEnvironmentStrippingVisitor;
 import org.quiltmc.loader.impl.util.FileSystemUtil;
 import org.quiltmc.loader.impl.util.FileUtil;
 import org.quiltmc.loader.impl.util.ManifestUtil;
-import org.quiltmc.loader.impl.util.PackageStrippingDataContainer;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
 import org.quiltmc.loader.impl.util.SystemProperties;
@@ -55,7 +52,6 @@ import java.nio.file.Path;
 import java.security.CodeSource;
 import java.security.cert.Certificate;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -326,17 +322,13 @@ class KnotClassDelegate {
 				// No package-info class file
 				return true;
 			}
-			PackageStrippingDataContainer data = new PackageStrippingDataContainer();
 
 			ClassReader reader = new ClassReader(bytes);
 
-			PackageEnvironmentStrippingVisitor envStripping = new PackageEnvironmentStrippingVisitor(QuiltLoaderImpl.ASM_VERSION, data, envType);
-			reader.accept(envStripping, ClassReader.SKIP_CODE | ClassReader.SKIP_FRAMES);
+			PackageStrippingData strippingData = new PackageStrippingData(QuiltLoaderImpl.ASM_VERSION, envType, modCodeSourceMap);
+			reader.accept(strippingData, ClassReader.SKIP_CODE | ClassReader.SKIP_FRAMES);
 
-			PackageRequiresStrippingVisitor requiresStripping = new PackageRequiresStrippingVisitor(QuiltLoaderImpl.ASM_VERSION, data, modCodeSourceMap);
-			reader.accept(requiresStripping, ClassReader.SKIP_CODE | ClassReader.SKIP_FRAMES);
-
-			return !data.stripEntirePackage();
+			return !strippingData.stripEntirePackage();
 		} catch (IOException e) {
 			throw new RuntimeException("Unable to load " + fileName, e);
 		}
