+++ src/main/java/org/quiltmc/loader/impl/transformer/EnvironmentStrippingData.java
@@ -1,0 +1,157 @@
+/*
+ * Copyright 2022, 2023 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.loader.impl.transformer;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+import org.objectweb.asm.AnnotationVisitor;
+import org.objectweb.asm.FieldVisitor;
+import org.objectweb.asm.MethodVisitor;
+import org.objectweb.asm.TypePath;
+import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
+import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
+
+import net.fabricmc.api.EnvType;
+
+/** Deprecated. All stuff were moved to {@link StrippingData}. */
+@Deprecated
+@QuiltLoaderInternal(QuiltLoaderInternalType.LEGACY_EXPOSED)
+public class EnvironmentStrippingData extends StrippingData {
+	// Fabric annotations
+	private static final String ENVIRONMENT_DESCRIPTOR = StrippingData.ENVIRONMENT_DESCRIPTOR;
+	private static final String ENVIRONMENT_INTERFACE_DESCRIPTOR = StrippingData.ENVIRONMENT_INTERFACE_DESCRIPTOR;
+	private static final String ENVIRONMENT_INTERFACES_DESCRIPTOR = StrippingData.ENVIRONMENT_INTERFACES_DESCRIPTOR;
+
+	// Quilt annotations
+	private static final String CLIENT_ONLY_DESCRIPTOR = StrippingData.CLIENT_ONLY_DESCRIPTOR;
+	private static final String SERVER_ONLY_DESCRIPTOR = StrippingData.SERVER_ONLY_DESCRIPTOR;
+
+	private class FabricEnvironmentAnnotationVisitor extends StrippingData.FabricEnvironmentAnnotationVisitor {
+
+		private FabricEnvironmentAnnotationVisitor(int api, Runnable onEnvMismatch) {
+			super(api, onEnvMismatch);
+		}
+
+		@Override
+		public void visitEnum(String name, String descriptor, String value) {
+			super.visitEnum(name, descriptor, value);
+		}
+	}
+
+	private class FabricEnvironmentInterfaceAnnotationVisitor extends StrippingData.FabricEnvironmentInterfaceAnnotationVisitor {
+
+		private FabricEnvironmentInterfaceAnnotationVisitor(int api) {
+			super(api);
+		}
+
+		@Override
+		public void visitEnum(String name, String descriptor, String value) {
+			super.visitEnum(name, descriptor, value);
+		}
+
+		@Override
+		public void visit(String name, Object value) {
+			super.visit(name, value);
+		}
+
+		@Override
+		public void visitEnd() {
+			super.visitEnd();
+		}
+	}
+
+	private class QuiltEnvironmentAnnotationVisitor extends StrippingData.QuiltEnvironmentAnnotationVisitor {
+
+		private QuiltEnvironmentAnnotationVisitor(int api, Runnable onEnvMismatch, Runnable onEnvMismatchLambdas) {
+			super(api, onEnvMismatch, onEnvMismatchLambdas);
+		}
+
+		@Override
+		public void visit(String name, Object value) {
+			super.visit(name, value);
+		}
+
+		@Override
+		public void visitEnd() {
+			super.visitEnd();
+		}
+	}
+
+	AnnotationVisitor visitMemberAnnotation(String descriptor, boolean visible, Runnable onEnvMismatch, Runnable onEnvMismatchLambdas) {
+		return super.visitMemberAnnotation(descriptor, visible, onEnvMismatch, onEnvMismatchLambdas);
+	}
+
+	public EnvironmentStrippingData(int api, EnvType envType) {
+		super(api, envType, new ArrayList<>());
+	}
+
+	@Override
+	public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
+		super.visit(version, access, name, signature, superName, interfaces);
+	}
+
+	@Override
+	public AnnotationVisitor visitAnnotation(String descriptor, boolean visible) {
+		return super.visitAnnotation(descriptor, visible);
+	}
+
+	@Override
+	public AnnotationVisitor visitTypeAnnotation(int typeRef, TypePath typePath, String descriptor, boolean visible) {
+		return super.visitTypeAnnotation(typeRef, typePath, descriptor, visible);
+	}
+
+	@Override
+	public FieldVisitor visitField(int access, String name, String descriptor, String signature, Object value) {
+		return super.visitField(access, name, descriptor, signature, value);
+	}
+
+	@Override
+	public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
+		return super.visitMethod(access, name, descriptor, signature, exceptions);
+	}
+
+	@Override
+	public boolean stripEntireClass() {
+		return super.stripEntireClass();
+	}
+
+	@Override
+	public Collection<String> getStripInterfaces() {
+		return super.getStripInterfaces();
+	}
+
+	@Override
+	public Collection<String> getStripFields() {
+		return super.getStripFields();
+	}
+
+	@Override
+	public Collection<String> getStripMethods() {
+		return super.getStripMethods();
+	}
+
+	@Override
+	public Collection<String> getStripMethodLambdas() {
+		return super.getStripMethodLambdas();
+	}
+
+	@Override
+	public boolean isEmpty() {
+		return super.isEmpty();
+	}
+}
