+++ src/main/java/org/quiltmc/loader/impl/transformer/TransformCacheGenerator.java
@@ -64,7 +64,6 @@ final class TransformCacheGenerator {
 					accessWidener,
 					name,
 					mod,
-					modList,
 					classBytes
 			);
 		});
